moved OperationOptimizer
